refactor: convert to pnpm workspace with client/server/shared packages
- Create pnpm-workspace.yaml with client, server, shared packages
- Split dependencies into per-package package.json files
- Move .storybook/, .babelrc, panda.config.ts, postcss.config.cjs to client/
- Move config.ts to server/config.ts, update all imports to @server/config
- Remove legacy external: ['geoip-country'] from esbuild config
- Update webpack babel-loader to use explicit configFile path
- Update test setup mocks to use correct config path
- Create tsconfig.json for client and shared packages
- Remove unused root deps (ts-loader, tsconfig-paths, webpack)
