--- .storybook/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "jsx": "react-jsx"
-  },
-  "include": [
-    "./**/*.ts",
-    "./**/*.tsx",
-    "../client/src/javascript/**/*.ts",
-    "../client/src/javascript/**/*.tsx",
-    "../shared/**/*.ts"
-  ]
-}
