+++ client/package.json
@@ -1,0 +1,88 @@
+{
+  "name": "@flood/client",
+  "version": "0.0.0",
+  "private": true,
+  "devDependencies": {
+    "@flood/shared": "workspace:*",
+    "@babel/core": "^7.29.0",
+    "@babel/preset-env": "^7.29.2",
+    "@babel/preset-react": "^7.28.5",
+    "@babel/preset-typescript": "^7.28.5",
+    "@dnd-kit/core": "^6.3.1",
+    "@dnd-kit/modifiers": "^6.0.1",
+    "@dnd-kit/sortable": "^7.0.2",
+    "@dnd-kit/utilities": "^3.2.2",
+    "@lingui/cli": "^5.9.5",
+    "@lingui/core": "^5.9.5",
+    "@lingui/format-json": "^5.9.5",
+    "@lingui/loader": "^5.9.5",
+    "@lingui/react": "^5.9.5",
+    "@pandacss/dev": "^1.9.1",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.17",
+    "@storybook/addon-onboarding": "^9.1.20",
+    "@storybook/addon-webpack5-compiler-babel": "^3.0.6",
+    "@storybook/builder-webpack5": "^9.1.20",
+    "@storybook/react-webpack5": "^9.1.20",
+    "@storybook/test-runner": "^0.23.0",
+    "@types/d3": "^7.4.3",
+    "@types/d3-array": "^3.2.2",
+    "@types/d3-scale": "^4.0.9",
+    "@types/d3-shape": "^3.1.8",
+    "@types/lodash": "^4.17.24",
+    "@types/react": "^18.3.28",
+    "@types/react-dom": "^18.3.7",
+    "@types/react-measure": "^2.0.12",
+    "@types/react-transition-group": "^4.4.12",
+    "@types/react-window": "^1.8.8",
+    "autoprefixer": "^10.4.27",
+    "axios": "^1.15.0",
+    "babel-loader": "^8.4.1",
+    "case-sensitive-paths-webpack-plugin": "2.4.0",
+    "classnames": "^2.5.1",
+    "css-loader": "^6.11.0",
+    "css-minimizer-webpack-plugin": "^4.2.2",
+    "d3-array": "^3.2.4",
+    "d3-scale": "^4.0.2",
+    "d3-selection": "^3.0.0",
+    "d3-shape": "^3.2.0",
+    "eslint-webpack-plugin": "^4.2.0",
+    "fast-json-patch": "^3.1.1",
+    "fast-sort": "^3.4.1",
+    "file-loader": "^6.2.0",
+    "get-user-locale": "^1.5.1",
+    "html-webpack-plugin": "^5.6.6",
+    "js-file-download": "^0.4.12",
+    "lodash": "^4.18.1",
+    "mini-css-extract-plugin": "^2.10.2",
+    "mobx": "^6.15.0",
+    "mobx-react-lite": "^4.1.1",
+    "overlayscrollbars": "^2.15.1",
+    "overlayscrollbars-react": "^0.5.6",
+    "postcss": "^8.5.9",
+    "postcss-loader": "^7.3.4",
+    "raw-loader": "^4.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-dropzone": "^14.4.1",
+    "react-markdown": "^9.1.0",
+    "react-measure": "^2.5.2",
+    "react-refresh": "^0.14.2",
+    "react-router": "^6.30.3",
+    "react-router-dom": "^6.30.3",
+    "react-transition-group": "^4.4.5",
+    "react-use": "^17.6.0",
+    "react-window": "^1.8.11",
+    "ress": "^4.0.0",
+    "sass": "^1.99.0",
+    "sass-loader": "^16.0.7",
+    "source-map-loader": "^4.0.2",
+    "storybook": "^9.1.20",
+    "style-loader": "^3.3.4",
+    "terser-webpack-plugin": "^5.4.0",
+    "webpack": "^5.106.0",
+    "webpack-dev-server": "^5.2.3",
+    "webpack-shebang-plugin": "^1.1.8",
+    "webpackbar": "^7.0.0",
+    "zod": "^3.25.76"
+  }
+}
