+++ client/postcss.config.cjs
@@ -1,0 +1,8 @@
+module.exports = {
+  plugins: {
+    '@pandacss/dev/postcss': {
+      configPath: './client/panda.config.ts',
+    },
+    autoprefixer: {},
+  },
+};
