+++ client/src/javascript/components/torrent-list/TorrentListRow.stories.tsx
@@ -2,10 +2,10 @@ import type {Meta, StoryObj} from '@storybook/react-webpack5';
 import {within, expect, waitFor, userEvent} from 'storybook/test';
 
 import TorrentListRow from './TorrentListRow';
-import {MOCK_TORRENT_STATES} from '../../../../../.storybook/mocks/_fixtures';
+import {MOCK_TORRENT_STATES} from '../../../../.storybook/mocks/_fixtures';
 import TorrentStore from '@client/stores/TorrentStore';
 import FloodActions from '@client/actions/FloodActions';
-import MockStateStore from '../../../../../.storybook/mocks/MockStateStore';
+import MockStateStore from '../../../../.storybook/mocks/MockStateStore';
 import type {TorrentProperties} from '@shared/types/Torrent';
 import {
   createMockMouseEvent,
