+++ client/src/javascript/routes/Overview.stories.tsx
@@ -9,7 +9,7 @@ import React from 'react';
 
 import Overview from './Overview';
 // Using the mock FloodActions from Storybook mocks
-import FloodActions from '../../../../.storybook/mocks/FloodActions';
+import FloodActions from '../../../.storybook/mocks/FloodActions';
 import TorrentFilterStore from '@client/stores/TorrentFilterStore';
 import AuthActions from '@client/actions/AuthActions';
 import {createMockMouseEvent, TEST_TIMEOUTS} from '../test-utils/storybook-helpers';
@@ -18,8 +18,8 @@ import {TorrentPriority} from '@shared/types/Torrent';
 import {TorrentStatus} from '@shared/constants/torrentStatusMap';
 
 // Import mocks using relative paths since they're not in the webpack alias
-import MockStateStore from '../../../../.storybook/mocks/MockStateStore';
-import {MOCK_FLOOD_SETTINGS, MOCK_NOTIFICATIONS, SIZE, TIME} from '../../../../.storybook/mocks/_fixtures';
+import MockStateStore from '../../../.storybook/mocks/MockStateStore';
+import {MOCK_FLOOD_SETTINGS, MOCK_NOTIFICATIONS, SIZE, TIME} from '../../../.storybook/mocks/_fixtures';
 
 const meta: Meta<typeof Overview> = {
   title: 'Routes/Overview',
