+++ package.json
@@ -43,7 +43,7 @@
     ]
   },
   "scripts": {
-    "build": "panda codegen && node scripts/build.mjs",
+    "build": "panda codegen --cwd client && node scripts/build.mjs",
     "build-pkg": "rm -rf dist && npm run build && npm run pkg",
     "pkg": "pkg . --out-path dist-pkg --no-bytecode --public --options=enable-source-maps",
     "format-source": "prettier -w .",
@@ -52,95 +52,28 @@
     "lint": "NODE_ENV=development eslint --max-warnings 0 .",
     "prepack": "rm -rf dist && npm run build",
     "start": "node --enable-source-maps --use_strict dist/index.js",
-    "start:development:client": "concurrently \"panda --watch\" \"node client/scripts/development.mjs\"",
+    "start:development:client": "concurrently \"panda --cwd client --watch\" \"node client/scripts/development.mjs\"",
     "start:development:server": "NODE_ENV=development tsx watch server/bin/start.ts",
     "start:test:rtorrent": "node scripts/testsetup.mjs",
     "test": "vitest run --coverage",
     "test:watch": "vitest --watch",
     "lint-staged": "lint-staged",
     "prepare": "husky",
-    "storybook": "storybook dev -p 6006",
-    "build-storybook": "storybook build",
+    "storybook": "storybook dev -p 6006 --config-dir client/.storybook",
+    "build-storybook": "storybook build --config-dir client/.storybook",
     "test-storybook": "test-storybook"
   },
   "devDependencies": {
-    "@babel/core": "^7.29.0",
     "@babel/eslint-parser": "^7.28.6",
-    "@babel/preset-env": "^7.29.2",
-    "@babel/preset-react": "^7.28.5",
-    "@babel/preset-typescript": "^7.28.5",
-    "@dnd-kit/core": "^6.3.1",
-    "@dnd-kit/modifiers": "^6.0.1",
-    "@dnd-kit/sortable": "^7.0.2",
-    "@dnd-kit/utilities": "^3.2.2",
     "@eslint/js": "^9.39.4",
-    "@fastify/compress": "^8.3.1",
-    "@fastify/cookie": "^10.0.1",
-    "@fastify/rate-limit": "^10.3.0",
-    "@fastify/send": "^4.1.0",
-    "@fastify/sse": "^0.4.0",
-    "@fastify/static": "^8.3.0",
-    "@fastify/swagger": "^9.7.0",
-    "@lingui/cli": "^5.9.5",
-    "@lingui/core": "^5.9.5",
-    "@lingui/format-json": "^5.9.5",
-    "@lingui/loader": "^5.9.5",
-    "@lingui/react": "^5.9.5",
-    "@maxmind/geoip2-node": "^5.0.0",
-    "@pandacss/dev": "^1.9.1",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.17",
-    "@seald-io/nedb": "^3.1.0",
-    "@storybook/addon-onboarding": "^9.1.20",
-    "@storybook/addon-webpack5-compiler-babel": "^3.0.6",
-    "@storybook/builder-webpack5": "^9.1.20",
-    "@storybook/react-webpack5": "^9.1.20",
-    "@storybook/test-runner": "^0.23.0",
-    "@types/bencode": "^2.0.4",
-    "@types/content-disposition": "^0.5.9",
-    "@types/create-torrent": "^6.1.1",
-    "@types/d3": "^7.4.3",
-    "@types/d3-array": "^3.2.2",
-    "@types/d3-scale": "^4.0.9",
-    "@types/d3-shape": "^3.1.8",
     "@types/fs-extra": "^9.0.13",
-    "@types/http-errors": "^1.8.2",
     "@types/jest": "^30.0.0",
-    "@types/jsonwebtoken": "^9.0.10",
-    "@types/lodash": "^4.17.24",
-    "@types/mime-types": "^3.0.1",
-    "@types/morgan": "^1.9.10",
     "@types/node": "^22.19.17",
-    "@types/parse-torrent": "^5.8.8",
-    "@types/passport": "^1.0.17",
-    "@types/passport-jwt": "^3.0.13",
-    "@types/react": "^18.3.28",
-    "@types/react-dom": "^18.3.7",
-    "@types/react-measure": "^2.0.12",
-    "@types/react-transition-group": "^4.4.12",
-    "@types/react-window": "^1.8.8",
-    "@types/supertest": "^2.0.16",
-    "@types/tar-fs": "^2.0.4",
-    "@types/yargs": "^17.0.35",
     "@typescript-eslint/eslint-plugin": "^8.58.1",
     "@typescript-eslint/parser": "^8.58.1",
     "@vitest/coverage-v8": "^2.1.9",
-    "autoprefixer": "^10.4.27",
-    "axios": "^1.15.0",
-    "axios-mock-adapter": "^2.1.0",
-    "babel-loader": "^8.4.1",
-    "bencode": "^2.0.3",
-    "case-sensitive-paths-webpack-plugin": "2.4.0",
     "chalk": "^4.1.2",
-    "classnames": "^2.5.1",
     "concurrently": "^9.2.1",
-    "content-disposition": "^0.5.4",
-    "create-torrent": "^6.1.0",
-    "css-loader": "^6.11.0",
-    "css-minimizer-webpack-plugin": "^4.2.2",
-    "d3-array": "^3.2.4",
-    "d3-scale": "^4.0.2",
-    "d3-selection": "^3.0.0",
-    "d3-shape": "^3.2.0",
     "esbuild": "^0.25.12",
     "eslint": "^9.39.4",
     "eslint-config-prettier": "^9.1.2",
@@ -149,79 +82,16 @@
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-simple-import-sort": "^12.1.1",
     "eslint-plugin-storybook": "^9.1.20",
-    "eslint-webpack-plugin": "^4.2.0",
-    "fast-json-patch": "^3.1.1",
-    "fast-sort": "^3.4.1",
-    "fastify": "^5.8.4",
-    "fastify-type-provider-zod": "^3.0.0",
-    "feedsub": "^0.7.8",
-    "file-loader": "^6.2.0",
-    "form-data": "^4.0.5",
     "fs-extra": "^10.1.0",
-    "get-user-locale": "^1.5.1",
     "globals": "^15.15.0",
-    "hash-wasm": "^4.12.0",
-    "html-webpack-plugin": "^5.6.6",
-    "http-errors": "^2.0.1",
     "husky": "^9.1.7",
-    "js-file-download": "^0.4.12",
-    "jsonwebtoken": "^9.0.3",
     "lint-staged": "^15.5.2",
-    "lodash": "^4.18.1",
-    "mime-types": "^3.0.2",
-    "mini-css-extract-plugin": "^2.10.2",
-    "mobx": "^6.15.0",
-    "mobx-react-lite": "^4.1.1",
-    "morgan": "^1.10.1",
-    "overlayscrollbars": "^2.15.1",
-    "overlayscrollbars-react": "^0.5.6",
-    "p-queue": "^7.4.1",
-    "parse-torrent": "^9.1.5",
-    "passport": "^0.6.0",
-    "passport-jwt": "^4.0.1",
-    "pino": "^10.3.1",
-    "pino-pretty": "^13.1.3",
-    "postcss": "^8.5.9",
-    "postcss-loader": "^7.3.4",
     "prettier": "^2.8.8",
-    "raw-loader": "^4.0.2",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-dropzone": "^14.4.1",
-    "react-markdown": "^9.1.0",
-    "react-measure": "^2.5.2",
-    "react-refresh": "^0.14.2",
-    "react-router": "^6.30.3",
-    "react-router-dom": "^6.30.3",
-    "react-transition-group": "^4.4.5",
-    "react-use": "^17.6.0",
-    "react-window": "^1.8.11",
-    "ress": "^4.0.0",
-    "sanitize-filename": "^1.6.4",
-    "sass": "^1.99.0",
-    "sass-loader": "^16.0.7",
-    "saxen": "^8.1.2",
-    "source-map-loader": "^4.0.2",
-    "storybook": "^9.1.20",
-    "style-loader": "^3.3.4",
-    "supertest": "^7.2.2",
-    "tar-fs": "^3.1.2",
-    "terser-webpack-plugin": "^5.4.0",
-    "tldts": "^5.7.112",
-    "ts-loader": "^9.5.7",
-    "tsconfig-paths": "^4.2.0",
     "tsx": "^4.21.0",
-    "typed-emitter": "^2.1.0",
     "typescript": "~5.0.4",
     "typescript-eslint": "^8.58.1",
     "vite-tsconfig-paths": "^5.1.4",
-    "vitest": "^2.1.9",
-    "webpack": "^5.106.0",
-    "webpack-dev-server": "^5.2.3",
-    "webpack-shebang-plugin": "^1.1.8",
-    "webpackbar": "^7.0.0",
-    "yargs": "^17.7.2",
-    "zod": "^3.25.76"
+    "vitest": "^2.1.9"
   },
   "engines": {
     "node": ">=22.0.0"
