+++ server/.jest/auth.setup.js
@@ -32,7 +32,7 @@ const config = {
   disableRateLimit: false,
 };
 
-vi.mock('../../config', () => ({
+vi.mock('../config', () => ({
   __esModule: true,
   default: config,
 }));
