+++ server/.jest/qbittorrent.setup.js
@@ -42,7 +42,7 @@ const config = {
   disableRateLimit: false,
 };
 
-vi.mock('../../config', () => ({
+vi.mock('../config', () => ({
   __esModule: true,
   default: config,
 }));
