+++ server/.jest/transmission.setup.js
@@ -43,7 +43,7 @@ const config = {
   disableRateLimit: false,
 };
 
-vi.mock('../../config', () => ({
+vi.mock('../config', () => ({
   __esModule: true,
   default: config,
 }));
