+++ server/package.json
@@ -1,0 +1,59 @@
+{
+  "name": "@flood/server",
+  "version": "0.0.0",
+  "private": true,
+  "devDependencies": {
+    "@flood/shared": "workspace:*",
+    "@fastify/compress": "^8.3.1",
+    "@fastify/cookie": "^10.0.1",
+    "@fastify/rate-limit": "^10.3.0",
+    "@fastify/send": "^4.1.0",
+    "@fastify/sse": "^0.4.0",
+    "@fastify/static": "^8.3.0",
+    "@fastify/swagger": "^9.7.0",
+    "@maxmind/geoip2-node": "^5.0.0",
+    "@seald-io/nedb": "^3.1.0",
+    "@types/bencode": "^2.0.4",
+    "@types/content-disposition": "^0.5.9",
+    "@types/create-torrent": "^6.1.1",
+    "@types/http-errors": "^1.8.2",
+    "@types/jsonwebtoken": "^9.0.10",
+    "@types/mime-types": "^3.0.1",
+    "@types/morgan": "^1.9.10",
+    "@types/parse-torrent": "^5.8.8",
+    "@types/passport": "^1.0.17",
+    "@types/passport-jwt": "^3.0.13",
+    "@types/supertest": "^2.0.16",
+    "@types/tar-fs": "^2.0.4",
+    "@types/yargs": "^17.0.35",
+    "axios": "^1.15.0",
+    "axios-mock-adapter": "^2.1.0",
+    "bencode": "^2.0.3",
+    "content-disposition": "^0.5.4",
+    "create-torrent": "^6.1.0",
+    "fast-json-patch": "^3.1.1",
+    "fastify": "^5.8.4",
+    "fastify-type-provider-zod": "^3.0.0",
+    "feedsub": "^0.7.8",
+    "form-data": "^4.0.5",
+    "hash-wasm": "^4.12.0",
+    "http-errors": "^2.0.1",
+    "jsonwebtoken": "^9.0.3",
+    "mime-types": "^3.0.2",
+    "morgan": "^1.10.1",
+    "p-queue": "^7.4.1",
+    "parse-torrent": "^9.1.5",
+    "passport": "^0.6.0",
+    "passport-jwt": "^4.0.1",
+    "pino": "^10.3.1",
+    "pino-pretty": "^13.1.3",
+    "sanitize-filename": "^1.6.4",
+    "saxen": "^8.1.2",
+    "supertest": "^7.2.2",
+    "tar-fs": "^3.1.2",
+    "tldts": "^5.7.112",
+    "typed-emitter": "^2.1.0",
+    "yargs": "^17.7.2",
+    "zod": "^3.25.76"
+  }
+}
