+++ server/util/diskUsageUtil.ts
@@ -1,8 +1,7 @@
 import {spawn, SpawnOptions} from 'node:child_process';
 
+import config from '@server/config';
 import type {Disk} from '@shared/types/DiskUsage';
-
-import config from '../../config';
 
 const spawnAsync = (cmd: string, args: string[], options: SpawnOptions, maxBuffer: number): Promise<string> =>
   new Promise((resolve, reject) => {
