+++ shared/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@flood/shared",
+  "version": "0.0.0",
+  "private": true,
+  "devDependencies": {
+    "fast-json-patch": "^3.1.1",
+    "zod": "^3.25.76"
+  }
+}
